This is the Kandinsky Linear Spreader V1 Program
This is the Kandinsky Linear Spreader Program.
It does not very much. It spreads only a Set of different stroke weighted Lines
across the Canvas. This Program is the basic Step for a Set of much larger Programs.
I hope you will like this Program even when it does not very much.